requirements.txt
flask
mysql-connector-python
openai
requests
beautifulsoup4
